recreate PR 651
