+++ cluster-provider-client/provider.go
@@ -20,14 +20,9 @@ import (
 	"context"
 	"errors"
 
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/watch"
-	"k8s.io/klog/v2"
-
 	clusterprovider "github.com/kcp-dev/edge-mc/cluster-provider-client/cluster"
 	kindprovider "github.com/kcp-dev/edge-mc/cluster-provider-client/kind"
 	v1alpha1apis "github.com/kcp-dev/edge-mc/pkg/apis/logicalcluster/v1alpha1"
-	edgeclient "github.com/kcp-dev/edge-mc/pkg/client/clientset/versioned"
 )
 
 // Each provider gets its own namespace named prefixNamespace+providerName
@@ -37,67 +32,8 @@ func GetNamespace(providerName string) string {
 	return prefixNamespace + providerName
 }
 
-func ProcessProviderWatchEvents(ctx context.Context, w clusterprovider.Watcher, clientset edgeclient.Interface, providerName string) {
-	logger := klog.FromContext(ctx)
-	for {
-		event, ok := <-w.ResultChan()
-		if !ok {
-			w.Stop()
-			logger.Info("stopping")
-			// TODO: return an error
-			return
-		}
-		listLogicalClusters, err := clientset.LogicalclusterV1alpha1().LogicalClusters(providerName).List(ctx, v1.ListOptions{})
-		if err != nil {
-			logger.Error(err, "")
-			// TODO: how do we handle failure?
-			return
-		}
-		switch event.Type {
-		case watch.Added:
-			// TODO: I am currently ignoring the possibility of the logical cluster object already existing
-			var found bool = false
-			for _, logicalCluster := range listLogicalClusters.Items {
-				if logicalCluster.Name == event.Name {
-					found = true
-					break
-				}
-			}
-			if !found {
-				logger.Info("Creating new LogicalCluster object", event.Name)
-				var eventLogicalCluster v1alpha1apis.LogicalCluster
-				eventLogicalCluster.Name = event.Name
-				eventLogicalCluster.Spec.ClusterProviderDescName = providerName
-				eventLogicalCluster.Spec.Managed = false
-				eventLogicalCluster.Status.Phase = "Initializing"
-				_, err = clientset.LogicalclusterV1alpha1().LogicalClusters(providerName).Create(ctx, &eventLogicalCluster, v1.CreateOptions{})
-				if err != nil {
-					logger.Error(err, "")
-					// TODO: how do we handle failure?
-					return
-				}
-			}
-		case watch.Deleted:
-			logger.Info("Deleting LogicalCluster object", event.Name)
-			err := clientset.LogicalclusterV1alpha1().LogicalClusters(providerName).Delete(ctx, event.Name, v1.DeleteOptions{})
-			if err != nil {
-				// TODO: If the logical cluster object does not exist, ignore the error.
-				logger.Error(err, "")
-				// TODO: how do we handle failure?
-				return
-			}
-
-		default:
-			// Unknown!
-			logger.Info("unknown event")
-			// TODO return an error or panic?
-		}
-	}
-}
-
-// ES: return and error and don't panic, move push to outside the case
+// TODO: this is termporary for stage 1. For stage 2 we expect to have a uniform interface for all informers.
 func NewProvider(ctx context.Context,
-	clientset edgeclient.Interface,
 	providerName string,
 	providerType v1alpha1apis.ClusterProviderType) (ProviderClient, error) {
 	var newProvider ProviderClient = nil
