+++ config/exports/apiresourceschema-logicalclusters.logicalcluster.kubestellar.io.yaml
@@ -2,7 +2,7 @@ apiVersion: apis.kcp.io/v1alpha1
 kind: APIResourceSchema
 metadata:
   creationTimestamp: null
-  name: v230622-286e246c.logicalclusters.logicalcluster.kubestellar.io
+  name: v230622-b4b7445f.logicalclusters.logicalcluster.kubestellar.io
 spec:
   group: logicalcluster.kubestellar.io
   names:
@@ -38,7 +38,7 @@ spec:
                 resource
               type: string
               x-kubernetes-validations:
-              - message: ProviderName is immutable
+              - message: ClusterProviderDescName is immutable
                 rule: self == oldSelf
             Managed:
               default: true
